MultiAssayExperiment wrapper for IntegratedLearner

This function offers a simple interface to the IntegratedLearner() from a MultiAssayExperiment object. Instead of wrangling the data, the function does it for the user and feeds it to IntegratedLearner().

IntegratedLearnerFromMAE(
  mae,
  experiment,
  assay.type,
  outcome.col,
  valid.col = NULL,
  ...
)

Arguments

mae

MultiAssayExperiment

experiment

Character vector or Integer vector. Specifies experiments from mae.

assay.type

Character vector. Specified the name of assay for each experiment. The length must be equal to experiment.

outcome.col

Character scalar. Specifies the name of column from colData that includes the outcome variable.

valid.col

Character scalar. Specifies the name of column from colData that includes the information on which sample belongs to validation set. The variable must be logical. (Default: NULL)

...

additional arguments passed to IntegratedLearner()

na.rm: Logical scalar. Should features with missing values be dropped? (Default: FALSE)

Value

A SuperLearner object containing the trained model fits.