Integrated machine learning for multi-omics prediction and classification

Performs integrated machine learning to predict a binary or continuous outcome based on two or more omics layers (views). The IntegratedLearner function takes a training set (Y, X1, X2,...,Xn) and returns the predicted values based on a validation set. It also performs V-fold nested cross-validation to estimate the prediction accuracy of various fusion algorithms. Three types of integration paradigms are supported: early, late, and intermediate. The software includes multiple ML models based on the SuperLearner R package as well as several data exploration capabilities and visualization modules in a unified estimation framework.

IntegratedLearner(
  feature_table,
  sample_metadata,
  feature_metadata,
  feature_table_valid = NULL,
  sample_metadata_valid = NULL,
  folds = 5,
  seed = 1234,
  base_learner = "SL.BART",
  base_screener = "All",
  meta_learner = "SL.nnls.auc",
  run_concat = TRUE,
  run_stacked = TRUE,
  verbose = FALSE,
  print_learner = TRUE,
  refit.stack = FALSE,
  family = gaussian(),
  ...
)

Arguments

feature_table

An R data frame containing multiview features (in rows) and samples (in columns). Column names of feature_metadata must match the row names of sample_metadata.

sample_metadata

An R data frame of metadata variables (in columns). Must have a column named subjectID describing per-subject unique identifiers. For longitudinal designs, this variable is expected to have non-unique values. Additionally, a column named Y must be present which is the outcome of interest (can be binary or continuous). Row names of sample_metadata must match the column names of feature_table.

feature_metadata

An R data frame of feature-specific metadata across views (in columns) and features (in rows). Must have a column named featureID describing per-feature unique identifiers. Additionally, a column named featureType should describe the corresponding source layers. Row names of feature_metadata must match the row names of feature_table.

feature_table_valid

Feature table from validation set for which prediction is desired. Must have the exact same structure as feature_table. If missing, uses feature_table for feature_table_valid.

sample_metadata_valid

Sample-specific metadata table from independent validation set when available. Must have the exact same structure as sample_metadata.

folds

How many folds in the V-fold nested cross-validation? Default is 10.

seed

Specify the arbitrary seed value for reproducibility. Default is 1234.

base_learner

Base learner for late fusion and early fusion. Check out the SuperLearner user manual for all available options. Default is `SL.BART`.

base_screener

Whether to screen variables before fitting base models? All means no screening which is the default. Check out the SuperLearner user manual for all available options.

meta_learner

Meta-learner for late fusion (stacked generalization). Defaults to `SL.nnls.auc`. Check out the SuperLearner user manual for all available options.

run_concat

Should early fusion be run? Default is TRUE. Uses the specified base_learner as the learning algorithm.

run_stacked

Should stacked model (late fusion) be run? Default is TRUE.

verbose

logical; TRUE for SuperLearner printing progress (helpful for debugging). Default is FALSE.

print_learner

logical; Should a detailed summary be printed? Default is TRUE.

refit.stack

logical; For late fusion, post-refit predictions on the entire data is returned if specified. Default is FALSE.

family

Currently allows `gaussian()` for continuous or `binomial()` for binary outcomes.

...

Additional arguments. Not used currently.

Value

A SuperLearner object containing the trained model fits.

Author

Himel Mallick, him4004@med.cornell.edu